Legg til nytt steg -> Lagsaksbehandlingsblankett (kjøres etter vent på status fra iverksetting og før ferdigstill) Refaktorering av Vurderingsservice : unngå dependency loop (Del denne i to => VurderingStegService.kt).
