+++ src/main/kotlin/no/nav/familie/ef/sak/service/steg/VentePåStatusFraIverksett.kt
@@ -2,7 +2,7 @@ package no.nav.familie.ef.sak.service.steg
 
 import no.nav.familie.ef.sak.iverksett.IverksettClient
 import no.nav.familie.ef.sak.repository.domain.Behandling
-import no.nav.familie.ef.sak.task.FerdigstillBehandlingTask
+import no.nav.familie.ef.sak.task.LagSaksbehandlingsblankettTask
 import no.nav.familie.kontrakter.ef.iverksett.IverksettStatus
 import no.nav.familie.prosessering.domene.TaskRepository
 import org.springframework.stereotype.Service
@@ -11,16 +11,16 @@ import org.springframework.stereotype.Service
 class VentePåStatusFraIverksett(private val iverksettClient: IverksettClient, private val taskRepository: TaskRepository): BehandlingSteg<Void?> {
 
     override fun utførSteg(behandling: Behandling, data: Void?) {
-       return iverksettClient.hentStatus(behandling.id).let {
+        return iverksettClient.hentStatus(behandling.id).let {
             when (it) {
-                IverksettStatus.OK -> opprettFerdigstillOppgave(behandling)
-                else -> throw error("Mottok status $it fra iverksett for behandlingId=${behandling.id}" )
+                IverksettStatus.OK -> opprettLagSaksbehandlingsblankettTask(behandling)
+                else -> throw error("Mottok status $it fra iverksett for behandlingId=${behandling.id}")
             }
         }
     }
 
-    fun opprettFerdigstillOppgave(behandling: Behandling) {
-        taskRepository.save(FerdigstillBehandlingTask.opprettTask(behandling))
+    fun opprettLagSaksbehandlingsblankettTask(behandling: Behandling) {
+        taskRepository.save(LagSaksbehandlingsblankettTask.opprettTask(behandling.id))
     }
 
     override fun stegType(): StegType {
